requirements.txt
python-telegram-bot
markdown
matplotlib
pytz
httplib2
google-api-python-client
oauth2client==3.0.0
pyyaml